Separando projeto em diferentes modulos pom e configurando
