+++ .gitignore
@@ -1,4 +1,4 @@
 /bin/
 /target/
 /out/logs/*.log
-/webDriver/*
+/Web-test/webDriver/*
