+++ FrameworkBase-Web-test/src/resources/teste.feature
@@ -1,0 +1,12 @@
+@Tag
+Feature: testes
+
+  Scenario: Teste de 1
+    Given dado que
+    When quando
+    Then entao
+
+  Scenario: Teste de 2
+    Given dado que dois
+    When quando dois
+    Then entao dois
