+++ FrameworkBase-jUnit-test/src/test/core/exception/InvalidCharException.java
@@ -1,0 +1,3 @@
+package test.core.exception;
+public class InvalidCharException {
+}
