--- src/test/core/exception/InvalidCharException.java
@@ -1,2 +1,0 @@
-package test.core.exception;public class InvalidCharException {
-}
